\name{pedagdata}
\alias{pedagdata}
\docType{data}
\title{
  Pedagogic data set obtained during a binary Q-method
}
\description{
This data set is a small artificial pedagogic data set.
}
\usage{data("pedagdata")}
\format{
An initial data frame of dimensions 8 x 20 containing the binary ratings on 8 stimuli provided by a panel of 20 participants. '1' stands for a 'representative of the concept' assessment, and '0' stands for a 'not representative of the concept' assessment. The initial data frame is supplemented by 3 covariates providing information about the participants (Gender, Age, Freq.use; the first 2 categorical and the last one continuous); and by 5 covariates providing information about the stimuli (Citrus.fruits, Vanilla, Wood, Lotus, Packaging; the first 4 continuous and the last one categorical).
}
\author{Margot Brard \email{margot.brard@agrocampus-ouest.fr}}

\examples{
data(pedagdata)
}

\keyword{datasets}
